Issues v0.1.0 Issues.CLI

Handle the command line parsing and the dispatch to the various functions that end up generating a table of the last n issues in a github project

Functions

decode_response(arg)

main(argv)

parse_args(argv)

argv can be -h or —help, which returns :nelp.

Otherwise it is a github user name, project name, and (optionally) the number of entries to format.

Return a tuple of { user, project, count }, or :help if help was given.

process(arg1)

sort_into_ascending_order(list_of_issues)