Issues.CLI (issues v0.1.0)

Handle the command line parsing and the dispatch to the various functions that end up generating a table of the last n issues in a github project.

Functions

args_to_internal_representation(arg1)

@spec args_to_internal_representation([String.t()]) ::
  {String.t(), String.t(), integer()} | :help

@spec args_to_internal_representation([String.t()]) ::
  {String.t(), String.t(), integer()} | :help

@spec args_to_internal_representation(any()) :: :help

Convert the parsed arguments to an internal representation.

Parameters

args: List of parsed arguments.

Examples

iex> Issues.CLI.args_to_internal_representation(["elixir-lang", "elixir", "5"])
{"elixir-lang", "elixir", 5}

iex> Issues.CLI.args_to_internal_representation(["elixir-lang", "elixir"])
{"elixir-lang", "elixir", 4}

iex> Issues.CLI.args_to_internal_representation(["-h"])
:help

decode_response(arg)

@spec decode_response({:ok, any()} | {:error, map()}) :: any()

@spec decode_response({:error, map()}) :: :ok

Decode the response from the Github API.

Parameters

response: Tuple of {:ok, body} or {:error, error}.

Examples

iex> Issues.CLI.decode_response({:ok, body})
body

iex> Issues.CLI.decode_response({:error, error})
:ok

last(list, count)

@spec last([map()], integer()) :: [map()]

Return the last count items from the list.

Parameters

list: List of items.

count: Number of items to return.

Returns

The last count items from the list.

main(argv)

@spec main([String.t()]) :: :ok

Run the CLI with the given arguments.

Parameters

argv: List of command line arguments.

Examples

iex> Issues.CLI.run(["elixir-lang", "elixir", "5"])
:ok

parse_args(argv)

@spec parse_args([String.t()]) :: {String.t(), String.t(), integer()} | :help

Parse the command line arguments.

argv can be -h or --help, which returns :help. Otherwise it is a github user name, project name, and (optionally) the number of entries to format. Return a tuple of { user, project, count }, or :help if help was given.

Parameters

argv: List of command line arguments.

Examples

iex> Issues.CLI.parse_args(["elixir-lang", "elixir", "5"])
{"elixir-lang", "elixir", 5}

iex> Issues.CLI.parse_args(["-h"])
:help

process(arg1)

@spec process({String.t(), String.t(), integer()} | :help) :: :ok

@spec process({String.t(), String.t(), integer()}) :: :ok

Process the parsed arguments.

Parameters

args: Tuple of {user, project, count} or :help.

Examples

iex> Issues.CLI.process(:help)
:ok

iex> Issues.CLI.process({"elixir-lang", "elixir", 5})
:ok

sort_into_descending_order(list_of_issues)

@spec sort_into_descending_order([map()]) :: [map()]

Sort the list of issues into descending order.

Parameters

list_of_issues: List of issues.

Returns

The list of issues sorted into descending order.