Issues.CLI (Issues v0.1.0)

Handle the command line parsing and the dispatch to the various functions that end up generating a table of the last n issues in a github project

Functions

args_to_internal_representation(arg1)

decode_response(arg)

last(list, count)

main(argv)

parse_args(argv)

argv can be -h or --help, which returns :help.

Otherwise it is a github user name, project name, and (optionally) the number of entries to format.

Return a tuple of { user, project, count }, or :help if help was given.

process(arg1)

sort_into_descending_order(list_of_issues)